DigitDialog and DigitToast is now an element after refactor
