+++ src/elements/digit-dialog/index.js
@@ -1,0 +1,3 @@
+import DigitDialog from "./DigitDialog.element";
+
+export default DigitDialog;
