+++ src/elements/digit-toast/index.js
@@ -1,0 +1,3 @@
+import DigitToast from "./DigitToast.element";
+
+export default DigitToast;
