--- src/views/digit-dialog/index.js
@@ -1,3 +1,0 @@
-import DigitDialog from "./DigitDialog.view";
-
-export default DigitDialog;
