--- src/views/digit-toast/index.js
@@ -1,3 +1,0 @@
-import DigitToast from "./DigitToast.view";
-
-export default DigitToast;
